added filter method to organize transactions by both month and year, organized RecentTransactions.jsx by seperating out part of file to DisplayTransactions.jsx, transactions are displaying correctly, still experiencing failure to update with new info - NS
